Split trybuild_staging into a module directory
Replace the oversized single file with `trybuild_staging/mod.rs` for the
production helpers and sibling `tests` / `prop_tests` modules so each
file stays within the line-count guideline without changing behaviour or
public API.
